Add command to install serverless cloud
Install serverless cloud as a global package to be able to proceed with the cloud initialization.

Without an installation there was the following error on my Mac machine:
/bin/sh: cloud: command not found

In (the) future might be nice to be able to use npx. With that, both steps are merged into one - more or less.
The best example for that optimization are classics like create-react-app to look at.